backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
psycopg[binary]>=3.1.12
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# External Services
replicate==0.22.0
httpx==0.26.0
openai==1.54.0

# Image Processing
pillow==10.4.0
requests==2.32.3

# WebSocket
websockets==12.0

# AWS S3
boto3==1.34.34

# CORS & Middleware
python-dateutil==2.8.2
